updated scripts

+++ uninstall.sh
@@ -1,6 +1,44 @@
 #!/bin/bash
 
+
+
+export DEBIAN_FRONTEND=noninteractive
+
+echo 
+
+echo "NGINX SIMPLE UNINSTALL SCRIPT" 
+
+echo 
+
+read -p 'ENTER DOMAIN NAME WITHOUT WWW PREFIX (eg: mm.example.com) : ' DOMAIN
+
+echo
+
+read -p "CONTINUE UNINSTALLING NGINX ON $DOMAIN ? (TYPE 'Y' TO CONTINUE) : " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+
+echo "STARTING UNINSTALLATION !"
+
+
+#remove nginx
+
 apt-get purge nginx nginx-common -y
-rm -rf install.sh
-rm -rf /var/www/*
+
+#remove all SSL certs
+
+rm -rf /etc/letsencrypt/live/$DOMAIN
+
+#remove the directory created
+
+rm -rf /var/www/$DOMAIN
+
+rm -- "$0"
+
+
+fi
+
+
+
